requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
logfire
google-genai
langchain
langchain-google-genai
langgraph
firecrawl
aiohttp